requirements.txt
Cython
progressbar
TensorBoard
tqdm
scipy
pycocotools
pillow
https://download.pytorch.org/whl/cu100/torch-1.1.0-cp36-cp36m-linux_x86_64.whl
https://download.pytorch.org/whl/cu100/torchvision-0.3.0-cp36-cp36m-linux_x86_64.whl
matplotlib
pandas
sklearn
albumentations
apex
torchsummary
future